[refactor] :truck: rename newWalletController to walletController
